Added functions to the utils module, finished SD
Added the functions `isPageMissing` and `getCreator` to the `.utils` module so that they can be used in the other modules without having to be redefined. Added the code that allows SD to post a message on the creator's TP to let them know their article/page has been tagged as deleted.