Fixed username fetching from DOM in Especial:PáginasNuevas
As DOM is structured in a different way in Especial:PáginasNuevas, when using the inline report and warning buttons the script would query the DOM one parent level too far, thus not using the right data to fill the username variable that both reports and warnings module use. This change makes an exception for that page and limits the inclusion of the buttons to certain pages.